Split test command host effect tests
