+++ docs/COMPLETION_AUDIT.md
@@ -3758,6 +3758,11 @@ and do not assume Phase 4 is ready without evidence.
   focused child modules under
   `tests/integration/cli_build/build_command_validation/`, preserving the
   same diagnostics and execution checks without one mixed parent file.
+- `zen test build.zen` host-effect tests now split env-read rejection coverage
+  into focused basic and ordering modules under
+  `tests/integration/cli_build/graph_validation_test_command_host_effects/env_reads/`,
+  preserving undeclared env-read and validation-order checks while keeping
+  file-read coverage wired separately from the parent module.
 - Effect checking, typed allocator semantics, actors in std integration,
   JSON/YAML IR boundaries, and broader build graph execution remain gated by
   `docs/V1_SPEC.md`.
