+++ tests/integration/cli_build/graph_validation_test_command_host_effects/env_reads.rs
@@ -1,0 +1,4 @@
+#[path = "env_reads/basic.rs"]
+mod basic;
+#[path = "env_reads/ordering.rs"]
+mod ordering;
